requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
anyascii==0.3.2
    # via wagtail
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-browser-reload
beautifulsoup4==4.12.3
    # via wagtail
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
coverage==7.5.3
    # via -r requirements.in
crispy-tailwind==1.0.3
    # via -r requirements.in
defusedxml==0.7.1
    # via willow
dj-database-url==2.2.0
    # via -r requirements.in
django==5.0
    # via
    #   -r requirements.in
    #   crispy-tailwind
    #   dj-database-url
    #   django-allauth
    #   django-browser-reload
    #   django-crispy-forms
    #   django-extensions
    #   django-filter
    #   django-mailjet
    #   django-modelcluster
    #   django-permissionedforms
    #   django-taggit
    #   django-template-partials
    #   django-treebeard
    #   djangorestframework
    #   laces
    #   wagtail
django-allauth==65.2.0
    # via -r requirements.in
django-browser-reload==1.16.0
    # via -r requirements.in
django-crispy-forms==2.4
    # via crispy-tailwind
django-extensions==4.1
    # via -r requirements.in
django-filter==24.3
    # via wagtail
django-mailjet==0.3.1
    # via -r requirements.in
django-modelcluster==6.4
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-taggit==6.1.0
    # via wagtail
django-template-partials==24.4
    # via -r requirements.in
django-treebeard==4.7.1
    # via wagtail
djangorestframework==3.16.0
    # via wagtail
draftjs-exporter==5.1.0
    # via wagtail
et-xmlfile==2.0.0
    # via openpyxl
filetype==1.2.0
    # via willow
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
l18n==2021.3
    # via wagtail
laces==0.1.2
    # via wagtail
mailjet-rest==1.3.4
    # via
    #   -r requirements.in
    #   django-mailjet
openpyxl==3.1.5
    # via wagtail
packaging==25.0
    # via pytest
pillow==11.2.1
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.21.0
    # via willow
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements.in
pytest==8.3.3
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-dotenv
pytest-django==4.8.0
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pytest-dotenv
pytz==2025.2
    # via l18n
requests==2.32.3
    # via
    #   mailjet-rest
    #   wagtail
six==1.17.0
    # via l18n
soupsieve==2.7
    # via beautifulsoup4
sqlparse==0.5.3
    # via django
telepath==0.3.1
    # via wagtail
typing-extensions==4.13.2
    # via dj-database-url
urllib3==2.4.0
    # via requests
wagtail==6.3.2
    # via -r requirements.in
whitenoise==6.8.2
    # via -r requirements.in
willow==1.10.0
    # via wagtail
